+++ src/commands/outputs/put-item.ts
@@ -1,6 +1,6 @@
 import type { PutCommandOutput } from "@aws-sdk/lib-dynamodb";
 import type { Attributes } from "../../types.js";
-import { maybeSet } from "./lib.js";
+import { ObjectBuilder } from "./lib.js";
 
 // TODO wrap in an object to support later upcoming types.
 export type PutItemOutput<T extends Attributes = Attributes> = { item?: T };
@@ -12,5 +12,7 @@ export const PutItemOutput = {
 function decode<T extends Attributes = Attributes>(
   output: PutCommandOutput,
 ): PutItemOutput<T> {
-  return maybeSet({}, "item", output, "Attributes");
+  return ObjectBuilder.of<PutItemOutput<T>>()
+    .setIfNotNull("item", output.Attributes)
+    .unwrap();
 }