+++ test/unit/commands/inputs/index.ts
@@ -1,0 +1,8 @@
+import { suite } from "node:test";
+import { getItemTests } from "./get-item.js";
+
+export const inputTests = () => {
+  suite("inputs", () => {
+    getItemTests();
+  });
+};